chore: reorganised content check dirs in z3-logic into new check-proof-content dir
