+++ src/utilities/handle-check-proof-validity/check-proof-content/check-law-proof-content.ts
@@ -1,0 +1,2 @@
+export {checkHskipLawProof} from '@/utilities/handle-check-proof-validity/check-proof-content/hskip-law-logic/check-hskip-law-proof';
+export {checkHassignLawProof} from '@/utilities/handle-check-proof-validity/check-proof-content/hassign-law-logic/check-hassign-law-proof';
