+++ src/utilities/handle-check-proof-validity/check-proof-content/hskip-law-logic/do-skip-law-checks.ts
@@ -1,4 +1,4 @@
-import {decomposeHoareLawLine} from '../common-funcs';
+import {decomposeHoareLawLine} from '../../z3-logic/common-funcs';
 import type {CollectedTripleProofLines} from '@/models/misc';
 import type {
 	OtherLawStructure,
