+++ src/utilities/handle-check-proof-validity/check-proof-content/subst-law-logic/do-subst-law-checks.ts
