+++ src/utilities/handle-check-proof-validity/check-proof-content/subst-law-logic/do-supporting-arith-law-checks.ts
