+++ src/utilities/handle-check-proof-validity/validity-check-utils/validate-hoare-triple-proof/decompose-proof-lines.ts
@@ -1,7 +1,7 @@
 import {
 	checkHskipLawProof,
 	checkHassignLawProof,
-} from '@/utilities/handle-check-proof-validity/z3-logic/hoare-law-z3-solver-funcs';
+} from '@/utilities/handle-check-proof-validity/check-proof-content/check-law-proof-content';
 import type {CollectedTripleProofLines} from '@/models/misc';
 
 async function decomposeProofLines(
