+++ src/utilities/handle-check-proof-validity/z3-logic/arith-law-z3-logic/add-variable-constraint.ts
