+++ src/utilities/handle-check-proof-validity/z3-logic/hassign-law-z3-logic/handle-equation-compose.ts
