--- src/utilities/handle-check-proof-validity/z3-logic/hoare-law-z3-solver-funcs.ts
@@ -1,2 +1,0 @@
-export {checkHskipLawProof} from '@/utilities/handle-check-proof-validity/z3-logic/hskip-law-logic/check-hskip-law-proof';
-export {checkHassignLawProof} from '@/utilities/handle-check-proof-validity/z3-logic/hassign-law-logic/check-hassign-law-proof';
